add make and docker files for server
