+++ server/Makefile
@@ -1,0 +1,4 @@
+VERSION = 0.0.1
+
+docker:
+	docker build --tag minesweeper:${VERSION} .
