--- server/api/__init__.py
@@ -1,3 +1,0 @@
-import api.requests
-import api.responses
-import api.service
