+++ server/minesweeper/__init__.py
@@ -1,2 +1,4 @@
+import minesweeper.api
 import minesweeper.minefield
 import minesweeper.cell
+
