+++ server/requirements.txt
@@ -1,2 +1,2 @@
-flask >= 2.0
-flask-cors >= 3.0
+flask == 2.0.3
+flask-cors == 3.0.10
